requirements.txt
beautifulsoup4==4.9.1
google-api-python-client==1.12.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
requests==2.24.0